Updated to the lastest upstream version
  to check s-yata/marisa-trie#2

+++ lib/marisa/grimoire/trie/louds-trie.cc
@@ -2,11 +2,11 @@
 #include <functional>
 #include <queue>
 
-#include "../algorithm.h"
-#include "header.h"
-#include "range.h"
-#include "state.h"
-#include "louds-trie.h"
+#include "marisa/grimoire/algorithm.h"
+#include "marisa/grimoire/trie/header.h"
+#include "marisa/grimoire/trie/range.h"
+#include "marisa/grimoire/trie/state.h"
+#include "marisa/grimoire/trie/louds-trie.h"
 
 namespace marisa {
 namespace grimoire {
@@ -301,7 +301,7 @@ void LoudsTrie::build_trie(Vector<T> &keys,
   }
 
   if (next_trie_.get() != NULL) {
-    config_.parse((next_trie_->num_tries() + 1) |
+    config_.parse(static_cast<int>((next_trie_->num_tries() + 1)) |
         next_trie_->tail_mode() | next_trie_->node_order());
   } else {
     config_.parse(1 | tail_.mode() | config.node_order() |
